+++ lib/open_weather/models/forecast.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require_relative 'forecast/hourly'
+require_relative 'forecast/forecast'
+require_relative 'forecast/city'